feat: Add weights to Random strategy
