+++ binancial/strategies/__init__.py
@@ -1,3 +1,3 @@
 from .mean_reversion_crossover import MeanReversionCrossover
-from .random import Random
+from .weighted_random import WeightedRandom
 from .volume_confirmed_momentum_crossover import VolumeConfirmedMomentumCrossover
